requirements.txt
tk >= 8.6
pycryptodomex >=3.16.0

